moved static inits to lazy-eval, fixed VueUtil.getJavaVersion bug
git-svn-id: https://tuftsvue.svn.sourceforge.net/svnroot/tuftsvue/trunk@884 d66e5487-c4fa-4dda-8c8e-ca23240363c0

+++ VUE2/src/tufts/vue/VueUtil.java
@@ -86,7 +86,13 @@ public class VueUtil
     }
 
 
-    public static double getJavaVersion()
+    /*
+     * Be sure to compare this value to a constant *float*
+     * value -- e.g. "1.4f" -- just "1.4" as double value may
+     * actually appear to be less than 1.4, as that can't
+     * be exactly represented by a double.
+     */
+    public static float getJavaVersion()
     {
         return javaVersion;
     }
